fix(revise): don't ask author to revise afrer they've just revised
Fixes a bug where if a manuscript receives the 'revise' verdict, the author is asked to revise; they revise and are immediately asked again to revise, and so on.
This arose because we've stopped populating the manuscript.decision property, which was used for determining the new version's status.
Instead we can simply use the manuscript.status property.